+++ apps/gophercon-korea-web-24/src/app/fonts.ts
@@ -1,0 +1,44 @@
+import localFont from "@next/font/local";
+
+const pretendard = localFont({
+  src: [
+    {
+      path: "../assets/fonts/Pretendard-Thin.woff2",
+      weight: "100",
+    },
+    {
+      path: "../assets/fonts/Pretendard-ExtraLight.woff2",
+      weight: "200",
+    },
+    {
+      path: "../assets/fonts/Pretendard-Light.woff2",
+      weight: "300",
+    },
+    {
+      path: "../assets/fonts/Pretendard-Regular.woff2",
+      weight: "400",
+    },
+    {
+      path: "../assets/fonts/Pretendard-Medium.woff2",
+      weight: "500",
+    },
+    {
+      path: "../assets/fonts/Pretendard-SemiBold.woff2",
+      weight: "600",
+    },
+    {
+      path: "../assets/fonts/Pretendard-Bold.woff2",
+      weight: "700",
+    },
+    {
+      path: "../assets/fonts/Pretendard-ExtraBold.woff2",
+      weight: "800",
+    },
+    {
+      path: "../assets/fonts/Pretendard-Black.woff2",
+      weight: "900",
+    },
+  ],
+});
+
+export { pretendard };
